Refactor(jira): Encapsulate duplicated code into base classes.
