+++ src/main/java/com/synopsys/integration/alert/channel/jira/cloud/JiraChannel.java
@@ -29,6 +29,8 @@ import org.springframework.stereotype.Component;
 
 import com.google.gson.Gson;
 import com.synopsys.integration.alert.channel.jira.cloud.util.JiraIssueHandler;
+import com.synopsys.integration.alert.channel.jira.cloud.util.JiraIssuePropertyHandler;
+import com.synopsys.integration.alert.channel.jira.cloud.util.JiraTransitionHandler;
 import com.synopsys.integration.alert.channel.jira.common.JiraConstants;
 import com.synopsys.integration.alert.channel.jira.common.JiraMessageParser;
 import com.synopsys.integration.alert.common.channel.issuetracker.IssueConfig;
@@ -86,8 +88,9 @@ public class JiraChannel extends IssueTrackerChannel {
         IssueService issueService = jiraCloudServiceFactory.createIssueService();
         IssuePropertyService issuePropertyService = jiraCloudServiceFactory.createIssuePropertyService();
         IssueSearchService issueSearchService = jiraCloudServiceFactory.createIssueSearchService();
-
-        JiraIssueHandler jiraIssueHandler = new JiraIssueHandler(issueService, issueSearchService, issuePropertyService, jiraProperties, jiraMessageParser, getGson());
+        JiraTransitionHandler jiraTransitionHandler = new JiraTransitionHandler(issueService);
+        JiraIssuePropertyHandler jiraIssuePropertyHandler = new JiraIssuePropertyHandler(issueSearchService, issuePropertyService);
+        JiraIssueHandler jiraIssueHandler = new JiraIssueHandler(issueService, jiraProperties, jiraMessageParser, getGson(), jiraTransitionHandler, jiraIssuePropertyHandler);
         return jiraIssueHandler.createOrUpdateIssues(jiraIssueConfig, content);
     }
