+++ src/main/java/com/synopsys/integration/alert/channel/jira/common/util/BaseJiraTransitionHandler.java
@@ -1,0 +1,89 @@
+package com.synopsys.integration.alert.channel.jira.common.util;
+
+import java.util.Optional;
+
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.synopsys.integration.alert.common.channel.issuetracker.IssueConfig;
+import com.synopsys.integration.alert.common.channel.issuetracker.IssueMissingTransitionException;
+import com.synopsys.integration.alert.common.channel.issuetracker.TransitionValidator;
+import com.synopsys.integration.alert.common.enumeration.ItemOperation;
+import com.synopsys.integration.exception.IntegrationException;
+import com.synopsys.integration.jira.common.model.components.IdComponent;
+import com.synopsys.integration.jira.common.model.components.StatusCategory;
+import com.synopsys.integration.jira.common.model.components.StatusDetailsComponent;
+import com.synopsys.integration.jira.common.model.components.TransitionComponent;
+
+public abstract class BaseJiraTransitionHandler implements TransitionValidator<TransitionComponent> {
+    public static final String TODO_STATUS_CATEGORY_KEY = "new";
+    public static final String DONE_STATUS_CATEGORY_KEY = "done";
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    protected abstract void performTransition(String issueKey, IdComponent transitionId) throws IntegrationException;
+
+    protected abstract StatusDetailsComponent getStatusDetails(String issueKey) throws IntegrationException;
+
+    @Override
+    public boolean doesTransitionToExpectedStatusCategory(TransitionComponent transition, String expectedStatusCategoryKey) {
+        StatusDetailsComponent statusDetails = transition.getTo();
+        StatusCategory statusCategory = statusDetails.getStatusCategory();
+        return StringUtils.equals(expectedStatusCategoryKey, statusCategory.getKey());
+    }
+
+    public boolean transitionIssueIfNecessary(String issueKey, IssueConfig jiraIssueConfig, ItemOperation operation) throws IntegrationException {
+        if (ItemOperation.UPDATE.equals(operation)) {
+            logger.debug("No transition required for this issue: {}.", issueKey);
+            return false;
+        }
+
+        Optional<String> transitionName = determineTransitionName(operation, jiraIssueConfig);
+        if (transitionName.isPresent()) {
+            boolean shouldAttemptTransition = isTransitionRequired(operation, getStatusDetails(issueKey));
+            if (shouldAttemptTransition) {
+                performTransition(issueKey, transitionName.get());
+                return true;
+            } else {
+                logger.debug("The issue {} is already in the status category that would result from this transition ({}).", issueKey, transitionName);
+            }
+        } else {
+            logger.debug("No transition name was provided so no transition will be performed for this operation: {}.", operation);
+        }
+        return false;
+    }
+
+    private boolean isTransitionRequired(ItemOperation operation, StatusDetailsComponent statusDetailsComponent) throws IntegrationException {
+        StatusCategory statusCategory = statusDetailsComponent.getStatusCategory();
+        if (ItemOperation.ADD.equals(operation)) {
+            // Should reopen?
+            return DONE_STATUS_CATEGORY_KEY.equals(statusCategory.getKey());
+        } else if (ItemOperation.DELETE.equals(operation)) {
+            // Should resolve?
+            return TODO_STATUS_CATEGORY_KEY.equals(statusCategory.getKey());
+        }
+        return false;
+    }
+
+    private void performTransition(String issueKey, String transitionName) throws IntegrationException {
+        logger.debug("Attempting the transition '{}' on the issue '{}'", transitionName, issueKey);
+        Optional<TransitionComponent> firstTransitionByName = retrieveIssueTransition(issueKey, transitionName);
+        if (firstTransitionByName.isPresent()) {
+            String transitionId = firstTransitionByName.map(TransitionComponent::getId).get();
+            performTransition(issueKey, new IdComponent(transitionId));
+        } else {
+            throw new IssueMissingTransitionException(issueKey, transitionName);
+        }
+    }
+
+    private Optional<String> determineTransitionName(ItemOperation operation, IssueConfig jiraIssueConfig) {
+        if (!ItemOperation.UPDATE.equals(operation)) {
+            if (ItemOperation.DELETE.equals(operation)) {
+                return jiraIssueConfig.getResolveTransition();
+            } else {
+                return jiraIssueConfig.getOpenTransition();
+            }
+        }
+        return Optional.empty();
+    }
+}
