+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/JiraServerChannel.java
@@ -31,6 +31,8 @@ import com.google.gson.Gson;
 import com.synopsys.integration.alert.channel.jira.common.JiraConstants;
 import com.synopsys.integration.alert.channel.jira.common.JiraMessageParser;
 import com.synopsys.integration.alert.channel.jira.server.util.JiraIssueHandler;
+import com.synopsys.integration.alert.channel.jira.server.util.JiraIssuePropertyHandler;
+import com.synopsys.integration.alert.channel.jira.server.util.JiraTransitionHandler;
 import com.synopsys.integration.alert.common.channel.issuetracker.IssueConfig;
 import com.synopsys.integration.alert.common.channel.issuetracker.IssueTrackerChannel;
 import com.synopsys.integration.alert.common.channel.issuetracker.message.IssueTrackerMessageResult;
@@ -85,8 +87,9 @@ public class JiraServerChannel extends IssueTrackerChannel {
         IssueService issueService = jiraServerServiceFactory.createIssueService();
         IssuePropertyService issuePropertyService = jiraServerServiceFactory.createIssuePropertyService();
         IssueSearchService issueSearchService = jiraServerServiceFactory.createIssueSearchService();
-
-        JiraIssueHandler jiraIssueHandler = new JiraIssueHandler(issueService, issueSearchService, issuePropertyService, jiraProperties, jiraMessageParser, getGson());
+        JiraTransitionHandler jiraTransitionHandler = new JiraTransitionHandler(issueService);
+        JiraIssuePropertyHandler jiraIssuePropertyHandler = new JiraIssuePropertyHandler(issueSearchService, issuePropertyService);
+        JiraIssueHandler jiraIssueHandler = new JiraIssueHandler(issueService, jiraProperties, jiraMessageParser, getGson(), jiraTransitionHandler, jiraIssuePropertyHandler);
         return jiraIssueHandler.createOrUpdateIssues(jiraIssueConfig, content);
     }
