+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/util/JiraIssueHandler.java
@@ -22,90 +22,45 @@
  */
 package com.synopsys.integration.alert.channel.jira.server.util;
 
-import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Optional;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
 import com.synopsys.integration.alert.channel.jira.common.JiraMessageParser;
+import com.synopsys.integration.alert.channel.jira.common.util.BaseJiraIssueHandler;
 import com.synopsys.integration.alert.channel.jira.server.JiraServerProperties;
-import com.synopsys.integration.alert.channel.jira.server.descriptor.JiraServerDescriptor;
-import com.synopsys.integration.alert.common.channel.issuetracker.IssueConfig;
-import com.synopsys.integration.alert.common.channel.issuetracker.IssueContentModel;
-import com.synopsys.integration.alert.common.channel.issuetracker.IssueHandler;
-import com.synopsys.integration.alert.common.enumeration.ItemOperation;
-import com.synopsys.integration.alert.common.exception.AlertException;
-import com.synopsys.integration.alert.common.exception.AlertFieldException;
 import com.synopsys.integration.alert.common.message.model.ComponentItem;
 import com.synopsys.integration.alert.common.message.model.LinkableItem;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.jira.common.model.request.IssueCommentRequestModel;
+import com.synopsys.integration.jira.common.model.request.builder.IssueRequestModelFieldsMapBuilder;
 import com.synopsys.integration.jira.common.model.response.IssueResponseModel;
-import com.synopsys.integration.jira.common.rest.service.IssuePropertyService;
-import com.synopsys.integration.jira.common.server.builder.IssueRequestModelFieldsBuilder;
 import com.synopsys.integration.jira.common.server.model.IssueCreationRequestModel;
 import com.synopsys.integration.jira.common.server.model.IssueSearchIssueComponent;
 import com.synopsys.integration.jira.common.server.model.IssueSearchResponseModel;
-import com.synopsys.integration.jira.common.server.service.IssueSearchService;
 import com.synopsys.integration.jira.common.server.service.IssueService;
-import com.synopsys.integration.rest.exception.IntegrationRestException;
 
-public class JiraIssueHandler extends IssueHandler<IssueResponseModel> {
-    public static final String DESCRIPTION_CONTINUED_TEXT = "(description continued...)";
-
+public class JiraIssueHandler extends BaseJiraIssueHandler {
     private static final Logger logger = LoggerFactory.getLogger(JiraIssueHandler.class);
 
     private final IssueService issueService;
     private final JiraServerProperties jiraProperties;
-    private final Gson gson;
 
-    private final JiraTransitionHandler jiraTransitionHelper;
     private final JiraIssuePropertyHandler jiraIssuePropertyHelper;
 
-    public JiraIssueHandler(IssueService issueService, IssueSearchService issueSearchService, IssuePropertyService issuePropertyService, JiraServerProperties jiraProperties, JiraMessageParser jiraMessageParser, Gson gson) {
-        super(jiraMessageParser);
+    public JiraIssueHandler(IssueService issueService, JiraServerProperties jiraProperties, JiraMessageParser jiraMessageParser, Gson gson, JiraTransitionHandler jiraTransitionHandler, JiraIssuePropertyHandler jiraIssuePropertyHandler) {
+        super(jiraMessageParser, gson, jiraTransitionHandler, jiraIssuePropertyHandler);
         this.issueService = issueService;
         this.jiraProperties = jiraProperties;
-        this.gson = gson;
-        this.jiraTransitionHelper = new JiraTransitionHandler(issueService);
-        this.jiraIssuePropertyHelper = new JiraIssuePropertyHandler(issueSearchService, issuePropertyService);
+        this.jiraIssuePropertyHelper = jiraIssuePropertyHandler;
     }
 
     @Override
-    protected IssueResponseModel createIssue(IssueConfig jiraIssueConfig, String providerName, LinkableItem topic, LinkableItem nullableSubTopic, ComponentItem arbitraryItem, String trackingKey, IssueContentModel contentModel)
-        throws IntegrationException {
-        IssueRequestModelFieldsBuilder fieldsBuilder = createFieldsBuilder(contentModel);
-        fieldsBuilder.setProject(jiraIssueConfig.getProjectId());
-        fieldsBuilder.setIssueType(jiraIssueConfig.getIssueType());
-        String issueCreator = jiraIssueConfig.getIssueCreator();
-
-        try {
-            IssueResponseModel issue = issueService.createIssue(new IssueCreationRequestModel(issueCreator, jiraIssueConfig.getIssueType(), jiraIssueConfig.getProjectName(), fieldsBuilder));
-            logger.debug("Created new Jira Cloud issue: {}", issue.getKey());
-            String issueKey = issue.getKey();
-            addIssueProperties(issueKey, providerName, topic, nullableSubTopic, arbitraryItem, trackingKey);
-            addComment(issueKey, "This issue was automatically created by Alert.");
-            for (String additionalComment : contentModel.getAdditionalComments()) {
-                String comment = String.format("%s \n %s", DESCRIPTION_CONTINUED_TEXT, additionalComment);
-                addComment(issueKey, comment);
-            }
-            return issue;
-        } catch (IntegrationRestException e) {
-            throw improveRestException(e, issueCreator);
-        }
-    }
-
-    @Override
-    protected boolean transitionIssue(IssueResponseModel issueModel, IssueConfig issueConfig, ItemOperation operation) throws IntegrationException {
-        return jiraTransitionHelper.transitionIssueIfNecessary(issueModel.getKey(), issueConfig, operation);
+    public IssueResponseModel createIssue(String issueCreator, String issueType, String projectName, IssueRequestModelFieldsMapBuilder fieldsBuilder) throws IntegrationException {
+        return issueService.createIssue(new IssueCreationRequestModel(issueCreator, issueType, projectName, fieldsBuilder));
     }
 
     @Override
@@ -136,15 +91,6 @@ public class JiraIssueHandler extends IssueHandler<IssueResponseModel> {
     }
 
     @Override
-    protected String createAdditionalTrackingKey(ComponentItem componentItem) {
-        if (!componentItem.collapseOnCategory()) {
-            LinkableItem categoryItem = componentItem.getCategoryItem();
-            return categoryItem.getName() + categoryItem.getValue();
-        }
-        return StringUtils.EMPTY;
-    }
-
-    @Override
     protected String getIssueKey(IssueResponseModel issueModel) {
         return issueModel.getKey();
     }
@@ -152,55 +98,6 @@ public class JiraIssueHandler extends IssueHandler<IssueResponseModel> {
     @Override
     protected String getIssueTrackerUrl() {
         return jiraProperties.getUrl();
-    }
-
-    private AlertException improveRestException(IntegrationRestException restException, String issueCreatorEmail) {
-        JsonObject responseContent = gson.fromJson(restException.getHttpResponseContent(), JsonObject.class);
-        List<String> responseErrors = new ArrayList<>();
-        if (null != responseContent) {
-            JsonObject errors = responseContent.get("errors").getAsJsonObject();
-            JsonElement reporterErrorMessage = errors.get("reporter");
-            if (null != reporterErrorMessage) {
-                return AlertFieldException.singleFieldError(
-                    JiraServerDescriptor.KEY_ISSUE_CREATOR, String.format("There was a problem assigning '%s' to the issue. Please ensure that the user is assigned to the project and has permission to transition issues.", issueCreatorEmail)
-                );
-            }
-
-            JsonArray errorMessages = responseContent.get("errorMessages").getAsJsonArray();
-            for (JsonElement errorMessage : errorMessages) {
-                responseErrors.add(errorMessage.getAsString());
-            }
-            responseErrors.add(errors.toString());
-        }
-
-        String message = restException.getMessage();
-        if (!responseErrors.isEmpty()) {
-            message += " | Details: " + StringUtils.join(responseErrors, ", ");
-        }
-
-        return new AlertException(message, restException);
-    }
-
-    private void addIssueProperties(String issueKey, String provider, LinkableItem topic, LinkableItem nullableSubTopic, ComponentItem componentItem, String alertIssueUniqueId) throws IntegrationException {
-        LinkableItem component = componentItem.getComponent();
-        Optional<LinkableItem> subComponent = componentItem.getSubComponent();
-
-        String subTopicName = nullableSubTopic != null ? nullableSubTopic.getName() : null;
-        String subTopicValue = nullableSubTopic != null ? nullableSubTopic.getValue() : null;
-
-        jiraIssuePropertyHelper.addPropertiesToIssue(issueKey, provider, topic.getName(), topic.getValue(), subTopicName, subTopicValue,
-            componentItem.getCategory(),
-            component.getName(), component.getValue(), subComponent.map(LinkableItem::getName).orElse(null), subComponent.map(LinkableItem::getValue).orElse(null),
-            alertIssueUniqueId
-        );
-    }
-
-    private IssueRequestModelFieldsBuilder createFieldsBuilder(IssueContentModel contentModel) {
-        IssueRequestModelFieldsBuilder fieldsBuilder = new IssueRequestModelFieldsBuilder();
-        fieldsBuilder.setSummary(contentModel.getTitle());
-        fieldsBuilder.setDescription(contentModel.getDescription());
-
-        return fieldsBuilder;
     }
 
 }
